Release 0.8.2 bump
I'm speaking with @ocramius and this is my proposal of release.
I found into the story the last commit with DoctrineModule 0.8.* and PHP >= 5.3
In my opinion we can tag in this point the 0.8.2 the last version before changeing in 0.9.*, security issues excluded.

After this commit the master branch support DoctrineModule 0.9.* and PHP >= 5.4